reorganize test code
